removed the blog/training parts. Folder reorganisation
