--- app/Models/Comment.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace App\Models;
-
-use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
-
-class Comment extends Model
-{
-    use HasFactory;
-    protected $fillable = ['body', 'post_id'];
-
-    public function post(){
-        return $this->belongsTo(Post::class);
-    }
-}
