+++ resources/views/dashboard/test.blade.php
