+++ resources/views/discussion/create.blade.php
@@ -1,4 +1,4 @@
-@extends('welcome')
+@extends('layouts.master')
 
 @section('content')
     <div class="home_content">
