+++ resources/views/layouts/icons.blade.php
