+++ resources/views/layouts/master.blade.php
@@ -1,30 +1,24 @@
-<!doctype html>
+<!DOCTYPE html>
 <html lang="en">
 
 <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <meta name="description" content="">
-    <meta name="author" content="">
-    <link rel="icon" href="/docs/4.0/assets/img/favicons/favicon.ico">
-    <link rel="canonical" href="https://getbootstrap.com/docs/4.0/examples/blog/">
-    <title>Blog Template for Bootstrap</title>
-    <link href="https://fonts.googleapis.com/css?family=Playfair+Display:700,900" rel="stylesheet">
-    <!-- Bootstrap core CSS -->
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css">
-    <link rel="stylesheet" type="text/css" href="{{ url('css/blog.css') }}">
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta name="description" content="This is my website constructed with Laravel">
+    <meta name="author" content="Guillaume Romero, Self taught">
+    <title>Ma page d'accueil</title>
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <link href='https://unpkg.com/boxicons@2.1.2/css/boxicons.min.css' rel='stylesheet'>
+    <link rel="stylesheet" type="text/css" href="{{ url('css/app.css') }}">
+    @include('layouts.icons')
 </head>
 
 <body>
-    <div class="container">
-        @include('layouts.nav')
+    <div class="content">
+        @include('layouts.sidebar')
+        @yield('content')
+        @yield('precision')
     </div>
-    <main role="main" class="container">
-        <div class="row">
-            @yield('content')
-        </div><!-- /.row -->
-    </main><!-- /.container -->
-    @include('layouts.footer')
 </body>
 
 </html>
