+++ resources/views/layouts/sidebar.blade.php
@@ -1,34 +1,125 @@
-<aside class="col-md-4 blog-sidebar">
-    <div class="p-3 mb-3 bg-light rounded">
-        <h4 class="font-italic">About</h4>
-        <p class="mb-0">Etiam porta <em>sem malesuada magna</em> mollis euismod. Cras mattis
-            consectetur purus sit amet fermentum. Aenean lacinia bibendum nulla sed consectetur.</p>
+<div class="sidebar">
+    <div class="logo_content">
+        <a href="/" class="logo">
+            <svg width="25" height="25">
+                <use xlink:href="#scorpion" />
+            </svg>
+            <span class="logo_name">Home</span>
+        </a>
+        <i class="bx bx-menu" id="btn"></i>
     </div>
+    <ul class="nav_list">
+        <li>
+            <i class='bx bx-search-alt-2' id="searchBtn"></i>
+            <input type="text" placeholder="Search...">
+            <span class="tooltip">Search</span>
+        </li>
+        <li>
+            <a href="/links">
+                <i class='bx bx-grid-alt'></i>
+                <span class="link_name">Dashboard</span>
+            </a>
+            <span class="tooltip">Dashboard</span>
+        </li>
+        <li>
+            <a href="#">
+                <i class='bx bx-user'></i>
+                <span class="link_name">User</span>
+            </a>
+            <span class="tooltip">User</span>
+        </li>
+        <li>
+            <a href="/discussions">
+                <i class='bx bxs-chat'></i>
+                <span class="link_name">Messages</span>
+            </a>
+            <span class="tooltip">Messages</span>
+        </li>
+        <li>
+            <a href="#">
+                <i class='bx bx-pie-chart-alt-2'></i>
+                <span class="link_name">Analitics</span>
+            </a>
+            <span class="tooltip">Analitics</span>
+        </li>
+        <li>
+            <a href="#">
+                <i class='bx bx-folder'></i>
+                <span class="link_name">File Manager</span>
+            </a>
+            <span class="tooltip">Files</span>
+        </li>
+        <li>
+            <a href="{{ route('settings') }}">
+                <i class='bx bx-cog'></i>
+                <span class="link_name">Setting</span>
+            </a>
+            <span class="tooltip">Setting</span>
+        </li>
+    </ul>
+    <div class="profile_content">
+        <div class="profile">
+            @auth
+                @if (Auth::user()->avatar_path == 'none')
+                    <img src="https://rcmi.fiu.edu/wp-content/uploads/sites/30/2018/02/no_user.png" alt="user_avatar"
+                        id="user_ico">
+                @else
+                    <img src="{{ Auth::user()->avatar_path }}" alt="user_avatar" id="user_ico">
+                @endif
+                <div class="profile_details">
+                    <div class="name">Romero Guillaume</div>
+                    <div class="job">Web Designer</div>
+                </div>
+                <div class="icons logoutbtn">
+                    <a href="{{ route('logout') }}"
+                        onclick="event.preventDefault(); document.getElementById('logout-form').submit();">
+                        <form id="logout-form" action="{{ route('logout') }}" method="POST" style="display: none;">
+                            @csrf
+                        </form>
+                        <i class='bx bx-log-out' id="logout"></i>
+                    </a>
+                </div>
+            @endauth
+            @guest
+                <img src="https://rcmi.fiu.edu/wp-content/uploads/sites/30/2018/02/no_user.png" alt="user_avatar"
+                    id="user_ico">
+                <div class="profile_details">
+                    <div class="name">Please log in</div>
+                    <div class="job"></div>
+                </div>
+                <div class="icons">
+                    <a href="{{ route('register') }}">
+                        <span class="tooltip">Register</span>
+                        <i class='bx bx-door-open' id="register"></i>
+                    </a>
+                    <a href="{{ route('login') }}">
+                        <span class="tooltip">Log in</span>
+                        <i class='bx bx-log-in' id="login"></i>
+                    </a>
+                </div>
+            @endguest
 
-    <div class="p-3">
-        <h4 class="font-italic">Archives</h4>
-        <ol class="list-unstyled mb-0">
-            <li><a href="#">March 2014</a></li>
-            <li><a href="#">February 2014</a></li>
-            <li><a href="#">January 2014</a></li>
-            <li><a href="#">December 2013</a></li>
-            <li><a href="#">November 2013</a></li>
-            <li><a href="#">October 2013</a></li>
-            <li><a href="#">September 2013</a></li>
-            <li><a href="#">August 2013</a></li>
-            <li><a href="#">July 2013</a></li>
-            <li><a href="#">June 2013</a></li>
-            <li><a href="#">May 2013</a></li>
-            <li><a href="#">April 2013</a></li>
-        </ol>
-    </div>
 
-    <div class="p-3">
-        <h4 class="font-italic">Elsewhere</h4>
-        <ol class="list-unstyled">
-            <li><a href="#">GitHub</a></li>
-            <li><a href="#">Twitter</a></li>
-            <li><a href="#">Facebook</a></li>
-        </ol>
+        </div>
     </div>
-</aside><!-- /.blog-sidebar -->
+</div>
+
+<script>
+    let btn = document.querySelector("#btn");
+    let searchBtn = document.querySelector("#searchBtn");
+    let user_icoBtn = document.querySelector("#user_ico");
+
+    let sidebar = document.querySelector(".sidebar");
+
+    btn.onclick = function() {
+        sidebar.parentElement.classList.toggle("sb_active");
+    }
+
+    searchBtn.onclick = function() {
+        sidebar.parentElement.classList.toggle("sb_active");
+    }
+
+    user_icoBtn.onclick = function() {
+        sidebar.parentElement.classList.toggle("sb_active");
+    }
+</script>
