--- resources/views/posts/post.blade.php
@@ -1,14 +1,0 @@
-<div class="blog-post">
-    <h2 class="blog-post-title">
-        <a href="/posts/{{ $post->id }}">
-            {{ $post->title }}
-        </a>
-    </h2>
-    <p class="blog-post-meta">
-        {{ $post->created_at->toFormattedDateString() }} by
-        <a href="/">Guillaume</a>
-    </p>
-    {{ $post->body }}
-    <br>
-    {{ $post->comments->count() }} comments
-</div>
