--- resources/views/site/layout/guest.blade.php
@@ -1,19 +1,0 @@
-@guest
-    <div class="no_home_content">
-        <svg>
-            <use xlink:href="#ouch" />
-        </svg>
-        <span>You are not connected</span>
-    </div>
-    <div class="no_precision">
-        <a href="/login">
-            <i class='bx bx-log-in'></i>
-            <span>Log In</span>
-        </a>
-        <hr>
-        <a href="/register">
-            <span>Sign In</span>
-            <i class='bx bx-door-open'></i>
-        </a>
-    </div>
-@endguest
